Consolidate Menu styles into token file
